Non working login, copy of the last of today

+++ Project/Web 2FA/lib/debug.php
@@ -1,5 +1,11 @@
 <?php
 
+if (FORCE_DEBUG_OFF === "false") 
+{
+	error_reporting(E_ALL);
+	ini_set("display_errors", true);
+}
+
 /*
 *
 * my_debug_print($value, __FILE__, __LINE__, "on");
@@ -9,40 +15,51 @@
 
 function my_debug_print($value, $file, $line, $debug = "off")
 {
-	if ($debug === "on") 
+	if (FORCE_DEBUG_OFF === "false") 
 	{
-		echo "<pre>";
-		echo "File: ".$file." Line: ".$line."<br>";
-		echo "Debug id: ".microtime(true)."<br>";
-		if (is_string($value)) 
+		if ($debug === "on") 
 		{
-			echo $value;
+			echo "<pre>";
+			echo "File: ".$file." Line: ".$line."<br>";
+			echo "Debug id: ".microtime(true)."<br>";
+			if (is_string($value)) 
+			{
+				echo $value;
+			}
+			else
+			{
+				var_dump($value);
+			}
+			echo "</pre>";
 		}
 		else
 		{
-			var_dump($value);
+			//Debug is off
 		}
-		echo "</pre>";
-	}
-	else
-	{
-		//Debug is off
 	}
 }
 
 function debug_string_backtrace() 
 { 
-    ob_start(); 
-    debug_print_backtrace(); 
-    $trace = ob_get_contents(); 
-    ob_end_clean(); 
+    if (FORCE_DEBUG_OFF === "false") 
+	{
 
-    // Remove first item from backtrace as it's this function which 
-    // is redundant. 
-    $trace = preg_replace ('/^#0\s+' . __FUNCTION__ . "[^\n]*\n/", '', $trace, 1); 
+	    ob_start(); 
+	    debug_print_backtrace(); 
+	    $trace = ob_get_contents(); 
+	    ob_end_clean(); 
 
-    // Renumber backtrace items. 
-    //$trace = preg_replace ('/^#(\d+)/me', '\'#\' . ($1 - 1)', $trace); //Do not work in php7
+	    // Remove first item from backtrace as it's this function which 
+	    // is redundant. 
+	    $trace = preg_replace ('/^#0\s+' . __FUNCTION__ . "[^\n]*\n/", '', $trace, 1); 
+
+	    // Renumber backtrace items. 
+	    //$trace = preg_replace ('/^#(\d+)/me', '\'#\' . ($1 - 1)', $trace); //Do not work in php7
+	}
+	else
+	{
+		$trace = "";
+	}
 
     return $trace; 
 }
